requirements.txt
absl-py==2.3.0
altair==5.5.0
attrs==25.3.0
bert-score==0.3.13
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
filelock==3.18.0
fonttools==4.58.4
fpdf==1.7.2
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
hf-xet==1.1.4
huggingface-hub==0.33.0
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
llvmlite==0.43.0
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.9.4
more-itertools==10.7.0
mpmath==1.3.0
narwhals==1.43.0
networkx==3.2.1
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
packaging==24.2
pandas==2.3.0
pillow==11.2.1
protobuf==6.31.1
pyarrow==20.0.0
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rpds-py==0.25.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
streamlit==1.45.1
summa==1.2.0
sympy==1.14.0
tenacity==9.1.2
textblob==0.19.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
zipp==3.23.0
